backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
passlib==1.7.4
bcrypt==4.2.1
python-jose==3.3.0
alembic==1.13.2
